Refaturando código e inserindo código de datas
