+++ main.py
@@ -1,56 +1,14 @@
-class ControleChassis:
-    def __init__(self, chassis, cliente, lote, mes, ano, observacoes):
-        self.chassis = chassis
-        self.cliente = cliente
-        self.lote = lote
-        self.mes = mes
-        self.ano = ano
-        self.observacoes = observacoes
+from datas_br import DatasBR
+from BuscaChassis import BuscaChassis
+from ControleChassis import ControleChassis
 
-    def cadastra(self):
-        arquivo_lista = open("registro_de_chassis.txt", "a", encoding="utf-8")
+cadastra_mes = DatasBR()
+formato_mes = cadastra_mes.mes_cadastro()
 
-        entrada_de_dados = f'Número de chassis = {self.chassis} - Nome do cliente = {self.cliente} - Lote = {self.lote} - Mês = {self.mes} - Ano = {self.ano} - Observações = {self.observacoes}'
-
-        arquivo_lista.write(entrada_de_dados)
-        arquivo_lista.write("\n")
-        arquivo_lista = open("registro_de_chassis.txt", "r", encoding="utf-8")
-        registro_completo = arquivo_lista.readlines()
-
-        cadastro = []
-
-        for linha in registro_completo:
-            linha = linha.strip()
-            cadastro.append(linha)
-
-        print(f'{self.cliente} cadastrado com sucesso')
-
-
-class BuscaChassis:
-    def __init__(self, busca):
-        self.busca = busca
-
-    def get_busca(self, busca):
-        arquivo_lista = open("registro_de_chassis.txt", "r", encoding="utf-8")
-        registro_completo = arquivo_lista.readlines()
-
-        cadastro = []
-        pesquisa_digitada = busca
-        for linha in registro_completo:
-            cadastro.append(linha)
-
-            elemento_encontrado = None
-
-            if pesquisa_digitada in linha:
-                elemento_encontrado = linha
-                print(elemento_encontrado)
-                break
-        if not elemento_encontrado:
-            print("Cadastro não encontrado")
-
+cadastra_ano = DatasBR()
+formato_ano = cadastra_ano.Ano()
 
 escolha = 0
-
 while escolha < 4:
 
     print("***************************************************")
@@ -68,8 +26,8 @@ while escolha < 4:
         numero_chassi = input("Digite o número de chassis\n")
         nome_cliente = input("Digite o nome do cliente\n")
         lote = input(f'Digite o lote do {nome_cliente}\n')
-        mes = input("Qual o mês que foi feito?\n")
-        ano = input("De qual ano?\n")
+        mes = formato_mes
+        ano = formato_ano
         observacoes = input("Tem alguma observação?\n")
 
         controle_2023 = ControleChassis(numero_chassi, nome_cliente, lote, mes, ano, observacoes)
